#33 Refactor DDD abstractions - move tolerance/time-adjustment logic from entity to aggregate - remove display color from domain enum and - fix self-instantiation in `before_save_worklog`
