+++ hr_time/tests/unit/test_worklog_entity.py
@@ -142,85 +142,3 @@ class TestWorklogEntity(unittest.TestCase):
             allocations=[]
         )
         self.assertEqual(0, entity.total_allocated)
-
-    def test_is_within_tolerance_true(self):
-        """Should return True when difference within tolerance"""
-        entity = WorklogEntity(
-            id=None,
-            employee_id="EMP-001",
-            log_time=datetime.now(),
-            work_desc="Test",
-            time_saved=3.5,
-            is_home_office=False,
-            ticket_link=None,
-            docstatus=DocumentStatus.DRAFT,
-            timesheet_id=None,
-            allocations=[self.alloc1, self.alloc2]
-        )
-        # total_allocated = 3.5, actual = 3.5, tolerance = 0.5
-        self.assertTrue(entity.is_within_tolerance(3.5, 0.5))
-
-    def test_is_within_tolerance_false(self):
-        """Should return False when difference exceeds tolerance"""
-        entity = WorklogEntity(
-            id=None,
-            employee_id="EMP-001",
-            log_time=datetime.now(),
-            work_desc="Test",
-            time_saved=3.5,
-            is_home_office=False,
-            ticket_link=None,
-            docstatus=DocumentStatus.DRAFT,
-            timesheet_id=None,
-            allocations=[self.alloc1, self.alloc2]
-        )
-        # total_allocated = 3.5, actual = 5.0, difference = 1.5 > 0.5
-        self.assertFalse(entity.is_within_tolerance(5.0, 0.5))
-
-    def test_adjusted_time_saved_overallocated(self):
-        """Should return total_allocated when allocated MORE than worked (within tolerance)"""
-        # Create allocations that total 5.0
-        alloc1 = TaskAllocation(
-            task_id="TASK-001", subject="Task 1",
-            time_spent=3.0, expected_time=8.0, progress_increment=37.5,
-            status="Working", priority="High"
-        )
-        alloc2 = TaskAllocation(
-            task_id="TASK-002", subject="Task 2",
-            time_spent=2.0, expected_time=4.0, progress_increment=50.0,
-            status="Working", priority="Medium"
-        )
-        entity = WorklogEntity(
-            id=None,
-            employee_id="EMP-001",
-            log_time=datetime.now(),
-            work_desc="Test",
-            time_saved=3.5,
-            is_home_office=False,
-            ticket_link=None,
-            docstatus=DocumentStatus.DRAFT,
-            timesheet_id=None,
-            allocations=[alloc1, alloc2]
-        )
-        # total_allocated = 5.0, actual = 3.5 (overallocated)
-        result = entity.adjusted_time_saved(3.5)
-        self.assertEqual(5.0, result)  # Returns total_allocated (stretch up)
-
-    def test_adjusted_time_saved_underallocated(self):
-        """Should return actual_work_hours when allocated LESS than worked"""
-        # Use the original allocations (total 3.5)
-        entity = WorklogEntity(
-            id=None,
-            employee_id="EMP-001",
-            log_time=datetime.now(),
-            work_desc="Test",
-            time_saved=3.5,
-            is_home_office=False,
-            ticket_link=None,
-            docstatus=DocumentStatus.DRAFT,
-            timesheet_id=None,
-            allocations=[self.alloc1, self.alloc2]  # total = 3.5
-        )
-        # total_allocated = 3.5, actual = 5.0 (underallocated)
-        result = entity.adjusted_time_saved(5.0)
-        self.assertEqual(5.0, result)  # Returns actual_work_hours
